pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel", "pip>=23"]
build-backend = "setuptools.build_meta"

[project]
name = "rsynccheck"
version = "0.0.1"
description = "Check the completeness of an rsync operation."
authors = [{name = "AYF", email = "realazthat@gmail.com"}]
license = {file = "LICENSE.md"}
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
  "Operating System :: OS Independent",
  'Programming Language :: Python :: 3.8',
  'Programming Language :: Python :: 3.9',
  'Programming Language :: Python :: 3.10',
  'Programming Language :: Python :: 3.11',
  'Programming Language :: Python :: 3.12'
]
dependencies = [
  "anyio >=4.3.0,<5",
  "humanize >=4,<5",
  "pathspec >=0.12,<1",
  "prettytable >=3,<4",
  "pydantic >=2,<3",
  "pydantic_core >=2,<3",
  "python-slugify >=8,<9",
  "pyyaml >=6,<7",
  "rich >=13,<14",
  "rich-argparse >=1,<2",
  "ruamel.yaml >=0.18,<0.19",
  "termcolor >=2,<3",
  "types-pyyaml >=6,<7",
  "typing-extensions >=4,<5"
]
keywords = [
]

[project.optional-dependencies]
# Set of dependencies that are known to work. To add a new dependency here, add
# a version range to the `projecct.dependencies` section above, and then run
# `EXTRA=prod bash scripts/pin-extra-reqs.sh`. Optionally, delete all the
# contents to repin all the dependencies (probably a good idea to do once in a
# while).
prod = [
  "annotated-types==0.7.0",
  "anyio==4.4.0",
  "exceptiongroup==1.2.1",
  "humanize==4.9.0",
  "idna==3.7",
  "markdown-it-py==3.0.0",
  "mdurl==0.1.2",
  "pathspec==0.12.1",
  "prettytable==3.10.0",
  "pydantic==2.7.3",
  "pydantic-core==2.18.4",
  "pygments==2.17.2",
  "python-slugify==8.0.4",
  "pyyaml==6.0.1",
  "rich==13.7.1",
  "rich-argparse==1.4.0",
  "ruamel-yaml==0.18.6",
  "ruamel-yaml-clib==0.2.8",
  "sniffio==1.3.1",
  "termcolor==2.4.0",
  "text-unidecode==1.3",
  "types-pyyaml==6.0.12.20240311",
  "typing-extensions==4.10.0",
  "wcwidth==0.2.13"
]
# Set of dev dependencies, all pinned, so that they are known to work. To add a
# new dependency here, add the unpinned package name here, and then run
# `EXTRA=dev bash scripts/pin-extra-reqs.sh`.
dev = [
  "annotated-types==0.7.0",
  "anyio==4.4.0",
  "argcomplete==3.2.3",
  "autoflake==2.3.1",
  "beautifulsoup4==4.12.3",
  "bs4==0.0.2",
  "certifi==2024.2.2",
  "cffi==1.16.0",
  "cfgv==3.4.0",
  "changeguard==0.3.1",
  "charset-normalizer==3.3.2",
  "colorama==0.4.6",
  "cryptography==42.0.5",
  "defusedxml==0.7.1",
  "distlib==0.3.8",
  "docutils==0.20.1",
  "exceptiongroup==1.2.1",
  "filelock==3.13.3",
  "humanize==4.9.0",
  "identify==2.5.35",
  "idna==3.6",
  "importlib-metadata==7.1.0",
  "importlib-resources==6.4.0",
  "isort==5.13.2",
  "jaraco-classes==3.3.1",
  "jaraco-context==4.3.0",
  "jaraco-functools==4.0.0",
  "jeepney==0.8.0",
  "jinja2==3.1.3",
  "keyring==25.0.0",
  "markdown-it-py==3.0.0",
  "markupsafe==2.1.5",
  "mdreftidy==0.3.0",
  "mdremotifier==0.3.1",
  "mdurl==0.1.2",
  "mistletoe==1.3.0",
  "more-itertools==10.2.0",
  "mypy==1.8.0",
  "mypy-extensions==1.0.0",
  "nh3==0.2.17",
  "nodeenv==1.8.0",
  "pathspec==0.12.1",
  "pexpect==4.9.0",
  "pip-licenses==4.3.4",
  "pkginfo==1.10.0",
  "platformdirs==4.2.0",
  "pre-commit==3.5.0",
  "prettytable==3.10.0",
  "ptyprocess==0.7.0",
  "pycparser==2.21",
  "pydantic==2.7.3",
  "pydantic-core==2.18.4",
  "pyflakes==3.2.0",
  "pygments==2.17.2",
  "pyright==1.1.352",
  "python-slugify==8.0.4",
  "pyyaml==6.0.1",
  "readme-renderer==43.0",
  "requests==2.31.0",
  "requests-toolbelt==1.0.0",
  "rfc3986==2.0.0",
  "rich==13.7.1",
  "rich-argparse==1.4.0",
  "ruamel-yaml==0.18.6",
  "ruamel-yaml-clib==0.2.8",
  "secretstorage==3.3.3",
  "sniffio==1.3.1",
  "snipinator==2.1.0",
  "soupsieve==2.5",
  "termcolor==2.4.0",
  "text-unidecode==1.3",
  "toml-sort==0.23.1",
  "tomli==2.0.1",
  "tomlkit==0.12.4",
  "twine==5.0.0",
  "types-beautifulsoup4==4.12.0.20240504",
  "types-colorama==0.4.15.20240311",
  "types-html5lib==1.1.11.20240228",
  "types-pyyaml==6.0.12.20240311",
  "typing-extensions==4.10.0",
  "urllib3==2.2.1",
  "virtualenv==20.25.1",
  "wcwidth==0.2.13",
  "xmltodict==0.13.0",
  "yapf==0.40.2",
  "yq==3.2.3",
  "zipp==3.18.1"
]

[project.scripts]
rsynccheck = "rsynccheck.cli:main"

[project.urls]
Homepage = "https://github.com/realazthat/rsynccheck"
Documentation = "https://github.com/realazthat/rsynccheck"
Repository = "https://github.com/realazthat/rsynccheck"

[tool.rsynccheck-project-metadata]
last_release = "0.0.1"
last_stable_release = "0.0.1"

[tool.setuptools]
packages = ["rsynccheck", "rsynccheck.utilities"]

[tool.tomlsort]
